+++ server/src/main/scala/chess/domain/GameState.scala
@@ -8,13 +8,13 @@ import chess.domain.Side.White
   * @param movesNow a [[Side]] which moves now.
   * @param board the current [[Chessboard]] state.
   * @param castlingsAvailable a [[List]] with all available castlings for both [[Side]]s.
-  * @param enPassantSquareOption an absence or a presence of an en passant [[Square]].
+  * @param enPassantSquareOption an absence or a presence of a coordinate with an en passant [[Square]].
   */
 final case class GameState(
     movesNow: Side,
     board: Chessboard,
     castlingsAvailable: List[Castling],
-    enPassantSquareOption: Option[Square]
+    enPassantSquareOption: Option[Coordinate]
 )
 
 /** A factory for [[GameState]] instances. */